Fix a conflict with UIAlertView categories.
Some external libraries use the same trick of DMAlert to provide blocks
instead of a delegate for UIAlertView completion.
When 2 libraries (for instance, UIKitCategoryAdditions and DMAlert)
using this same trick are used in the same project, one of the category
will not work anymore.

This patch fixes the issue by creating a sub-class of UIAlertView
instead of creating a category.